[qca-backport] fix AP powersave TX vs. wakeup race
Merged the patch from https://git.kernel.org/cgit/linux/kernel/git/jberg/\
mac80211.git/commit/?id=1d147bfa64293b2723c4fec50922168658e613ba
So do not push the patch to upstream.

There is a race between the TX path and the STA wakeup: while
a station is sleeping, mac80211 buffers frames until it wakes
up, then the frames are transmitted. However, the RX and TX
path are concurrent, so the packet indicating wakeup can be
processed while a packet is being transmitted.

This can lead to a situation where the buffered frames list
is emptied on the one side, while a frame is being added on
the other side, as the station is still seen as sleeping in
the TX path.

As a result, the newly added frame will not be send anytime
soon. It might be sent much later (and out of order) when the
station goes to sleep and wakes up the next time.

Additionally, it can lead to the crash.

Fix all this by synchronising both paths with a new lock.
Both path are not fastpath since they handle PS situations.

In a later patch we'll remove the extra skb queue locks to
reduce locking overhead.

Change-Id: Id78ac134b7a28c2b492f3e291ff9802524ca67e3
Signed-off-by: Miaoqing Pan <miaoqing@codeaurora.org>