[qca-backports] fix invalid descriptor discarding
Only set sc->rx.discard_next to rx_stats->rs_more when actually
discarding the current descriptor.

Also, fix a detection of broken descriptors:
First the code checks if the current descriptor is not done.
Then it checks if the next descriptor is done.
Add a check that afterwards checks the first descriptor again, because
it might have been completed in the mean time.

Change-Id: I0c3a94c75d03811f50aec9af0d706b73337b5b6a
Signed-off-by: Rajkumar Manoharan <rmanohar@codeaurora.org>