qca-backports: update kernel panic fix
Based on GFP flag given to DMA coherent allocation, the behaviour of
dma_free_coherent is changed. This behavioural diffrence is noticeable
in ARM platform. If DMA memory is allocated with GFP_KERNEL, free
coherent can not be called inside spin lock. This is causing kernel
crash in ARM platforms. Fix this by changing GFP flag to atomic.

Change-Id: I504fe795448b10fc831de7c308848d8346bb659a
Signed-off-by: Rajkumar Manoharan <rmanohar@codeaurora.org>