qca-backports: revert qca-backports packages name to mac80211
We recently added the support for overrides, which allows us to replace
OpenWrt core packages by putting an equivalent in feeds with the same
name.

So we're renaming the (previously called) qca-backports package into
mac80211, and we're reverting the name of all the packages inside to
their original name. This minimize the changes done to openwrt.git and
keep our specificities in feeds.

Change-Id: Iec65d640d0c56044771c8851ad223100ef09e6f7
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>