mac80211:  Fix interrupt storm
Promiscuous mode is enabled when wlan interface is added to
bridge. ath10k creates a monitor mode when promiscuous mode
is enabled. When monitor vdev is runing along with other
vdev(s) there is a huge number of interrupts generated
especially in noisy condition. Fix this by not enabling
promiscuous(monitor) mode when already a vdev is running.
This does not change the support of virtual interface of
type monitor along with other vdevs.

Change-Id: I8beace6c67c31c69b634888e0ab92344919fd436
Signed-off-by: Vasanthakumar Thiagarajan <vthiagar@codeaurora.org>