mac80211: Increase SKID length for 10.X FW
The current SKID length configuration causes firmware
to reject peer creation for not able to allocate
AST entries for peers. This issue is observed when
least significant 3 bytes are used ramdomly to create
client MAC addresses.

AST table SKID length configuration is increased to
maximum value to fix this issue.

Change-Id: I82231aa0c5718b5ddbcc1efcbd97db3f4dcf08c9
Signed-off-by: SenthilKumar Jegadeesan <sjegadee@codeaurora.org>