ath10k: fix quiet duration configuration
Duty cycle is nothing but percentage of one period in which a
signal is active. But it is wrongly interpreted as ratio of quiet
duration. So fix the quiet duration to the remaining percentage
of duty cycle (active ratio). Add support to configure quiet period
via sysfs entry. This will be helpful to experiment different quiet
period values along with different duty cycle ratio.

To configure quiet period as 30ms,

    echo 30 >/sys/class/ieee80211/phyX/device/quiet_period

Change-Id: I3bb5f240116a3750e0ee4de2aac6043f059f2261
Signed-off-by: Rajkumar Manoharan <rmanohar@codeaurora.org>