backports: fix survey reporting
This patch is from upstream,
c94aa7ef4647b81bc110300e0a1be777961d7a3a.

Number of channels is stored in a separate macro
in a header file and channel list is constructed
independently. The macro is used to define survey
array.

This fixes a recent regression introduced after
adding support for 144 channel. The regression
would lead to a warning and incomplete survey data
on channel 165:

  chan info: invalid frequency 5825 (idx 38 out of bounds)

Also make sure to enforce the sizes and avoid this
kind of problem in the future.

Change-Id: I7699f5722cfaa4ff69c154bdaeee0272300f80f0
Signed-off-by: Peter Oh <poh@codeaurora.org>