ath10k: Add the module parameter to control the BT-coex feature
Add the module parameter bt_coex to control the BT-coex feature.
As the BT-coex capability has to be enable at init time, so this
parameter is readonly and can't be changed at running time. The
default value of this parameter is false to avoid side effect.
For the FW doesn't supply BT-coex services at WMI service map.
This parameter will be ignored.

Change-Id: I46ac7a239eff2ab147132a51a54a4163be864523
Signed-off-by: Yanbo Li <yanbo@codeaurora.org>