mac80211: Fix smart antenna performance training logic
A bandwidth is considered active if we see the number of ppdus seen in the
bandwidth is above the corresponding ppdu threshold. Currently, we are using a
wrong threshold value (344) which is nearly 30 times higher than the expected
threshold values. To be noted there the distinct threshold values for each
bandwidth.

Also fix the data type of couple of variables (thrshld and pdelta) as there
are expected to have much larger values.

Change-Id: I1bf87ad341055e58144b9e606e6fad076fcfe93a
Signed-off-by: Ashok Raj Nagarajan <arnagara@codeaurora.org>