mac80211: Ask FW to send enough training packets
Hosts requests FW to send a configured number (640) packets to be trained. This
number is also the minimum number of train packets Host expects out of a
configured antenna during training. There are occassions when some packets are
missed or maligned and are not counted by the hosts. During these occasions, FW
might think it have send enough train packets to host while the host waits for
more train packets from the FW. As a result, training is never stopped.

Fix this by configuring the number of packets to be trained by FW to a higher
number and host will stop training when the required number of train packets
are received.

Change-Id: Ic281cecf71edcc9cfb5df0b585e84a5ba3bf1c68
Signed-off-by: Ashok Raj Nagarajan <arnagara@codeaurora.org>