mac80211: Fix Smart antenna secondary metric
During training, Smart Antenna checks the stats of different antennas with the
currently selected antenna. Rates and PER are the primary metric used for the
comparision. When there is a tie, SA uses their RSSI samples (10 samples for
each chain) as the tie breaker.

A counter is used for each of these two antennas that gets incremented every
time when their RSSI samples is better than the other. Towards the end, the
antenna which has maximum counter is the best antenna.

Currently there is a bug where the least better antenna is selected as the part
of this tie breaker. So fix this.

Change-Id: I1798d7859b5b4ca12cc42a63e885882e0c98b908
Signed-off-by: Ashok Raj Nagarajan <arnagara@codeaurora.org>