mac80211: Fix Smart Antenna bitrate calculation
Smart antenna algorithm uses cfg80211's bitrate calculation. The way cfg80211
calculates bitrate for HT and VHT is different. Incase of HT, cfg80211 will
itself determine the number of streams (nss) whereas in case of VHT, it expects
a nss value.

Currently, ath10k driver is providing same set of rate information to cfg80211
for both cases VHT and HT and hence, bitrate is calculated wrongly when
operating in a HT mode. As a result, Antenna selection based on performance
triggering is not happening correctly.

Fix this by seperating the filling of rate_info (mcs values and nss) based on
VHT or HT.

Change-Id: I2304075f01874319f141b6f3fca9c666118549e6
Signed-off-by: Ashok Raj Nagarajan <arnagara@codeaurora.org>
Signed-off-by: Maharaja Kennadyrajan <mkenna@codeaurora.org>