ath10k: add support for qca99x0 Rx descriptors
QCA99X0 chip has an extra 4 bytes in rx_msdu_start,
20 bytes in rx_msdu_end and 20 bytes in rx_ppdu_end structure
which are used in htt_rx_desc and HTT Rx ring offset setup.
This is necessary for correct Rx for QCA99X0 or Rx descriptors
will be overwritten and corrupted.

With this patch QCA988X and QCA6174 will have extra 44 bytes
padding in Rx descriptor layout which is harmless.

Change-Id: Ia6fbba9210a68a74d022598024854cea9b98622b
Signed-off-by: Peter Oh <poh@codeaurora.org>