mac80211: Upgrade package to 20150702
Backport is generated out of 4.1.0-rc6 kernel of ath-qca branch of
ath.git. The upstream merged patches are removed and existing patches
are reordered to fix repeated numbering.

This update also has the changes for QCA99X0 firmware boot and
wmi.

ath10k-firware is now pointing to firmware-5.bin as latest driver
looks for it. PCI SOC powersave change is causing failures in ARM
platform while bringing up AP in security mode. Until it is properly
root caused, disabling the feature for QCA9888.

Change-Id: Id998643db34032db08ca299e46a6e01150789178
Signed-off-by: Vasanthakumar Thiagarajan <vthiagar@codeaurora.org>