mac80211: Fix spurious wmi train commands after SA training completed
FW is not aware of when a training should start and when it should end. It will
send the train packets per host's request. Currently, after the first training,
host is not conveying the training completion information to FW. As a result,
FW would keep sending us the train packet and inturn we are generating a
spurious wmi command configuring train information with large num_pkts.

Fix this, by sending the wmi command configuring train information with
num_pkts to 0. Now the FW knows that it need not send any more train packets.

Change-Id: I4f901e1382a7571481b99d35d93862bee82a5024
Signed-off-by: Ashok Raj Nagarajan <arnagara@codeaurora.org>